fix: api 조회 성공하게 수정

+++ src/main/java/com/sprint/monew/domain/comment/dto/CommentDto.java
@@ -13,7 +13,7 @@ public record CommentDto(
         UUID userId,
         String userNickname,
         String content,
-        long likeCount,
+        int likeCount,
         boolean likedByMe,
         Instant createdAt
 ) {
