+++ src/main/java/com/sprint/monew/domain/interest/dto/SubscriptionDto.java
@@ -10,7 +10,7 @@ public record SubscriptionDto(
     UUID interestId,
     String interestName,
     List<String> interestKeywords,
-    int interestSubscriberCount,
+    long interestSubscriberCount,
     Instant createdAt
 ) {
